weather-app/backend/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.22.0
pandas==2.2.3
requests-cache==0.9.8
openmeteo-requests==0.0.9
retry-requests==0.3.0
python-dotenv==1.0.0
